:sparkles: Feat: 카테고리 api
Feat: 카테고리 api

+++ src/main/java/pickRAP/server/service/category/CategoryService.java
@@ -74,6 +74,10 @@ public class CategoryService {
 
     @Transactional
     public void delete(Long id) {
+        if(categoryRepository.findById(id).isEmpty()) {
+            throw new BaseException(BaseExceptionStatus.DONT_EXIST_CATEGORY);
+        }
+
         categoryRepository.deleteById(id);
     }
 }
